+++ src/services/htmlLinkedEditing.ts
@@ -3,10 +3,10 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
-import { TextDocument, Position, Range } from 'vscode-languageserver-types';
+import { TextDocument, Position, Range } from '../htmlLanguageTypes';
 import { HTMLDocument } from '../parser/htmlParser';
 
-export function findOnTypeRenameRanges(
+export function findLinkedEditingRanges(
   document: TextDocument,
   position: Position,
   htmlDocument: HTMLDocument